// sitat-fe-workspace/projects/app-programmi/src/app/modules/models/interventi/interventi.model.ts
export interface ExportInterventiAquistiForm { 
    contri?: number;
    searchForm?: InterventiSearchForm;
    showAcquistoRicompreso?: boolean;
    showAmbitoGeografico?: boolean;
    showAnnualita?: boolean;
    showApportoCapitalePrivato?: boolean;
    showAusa?: boolean;
    showCUI?: boolean;
    showCategoria?: boolean;
    showCodCuiLavoroCollegato?: boolean;
    showCodInt?: boolean;
    showCodIstat?: boolean;
    showContrattoEssere?: boolean;
    showCpv?: boolean;
    showCup?: boolean;
    showDescrizione?: boolean;
    showDurataContratto?: boolean;
    showImmobili?: boolean;
    showImportiAnnualita?: boolean;
    showLavoroC?: boolean;
    showLottoF?: boolean;
    showNuts?: boolean;
    showPriorita?: boolean;
    showRup?: boolean;
    showScadenza?: boolean;
    showSettore?: boolean;
    showTipologia?: boolean;
    showVariato?: boolean;
}

export interface ExportInterventiAquistiNonRipropostiForm { 
    contri?: number;
    showCui?: boolean;
    showCup?: boolean;
    showDescrizione?: boolean;
    showImporto?: boolean;
    showMotivo?: boolean;
    showPriorita?: boolean;
}

export interface InterventiSearchForm { 
    annualita?: string;
    codInterno?: string;
    codiceCUP?: string;
    descrizione?: string;
    idProgramma?: number;
    importoTotaleDa?: number;
    importoTotaleA?: number;
    numeroCui?: string;
    rup?: string;
}